backend/app/services/extractor.py
from pdfminer.high_level import extract_text

def extract_text_from_pdf(file_path):
    try:
        return extract_text(file_path)
    except Exception as e:
        print("PDF extraction error:", e)
        return ""
