backend/requirement.txt
flask
flask-cors
elasticsearch
pymongo
pdfminer.six
scikit-learn
requests
beautifulsoup4
